zatca

+++ database/migrations/2024_09_18_113727_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
             $table->string('name');
             $table->boolean('is_deleted')->default(false);
-            $table->double('price')->default(0.0);
+            $table->decimal('price',15,2)->default(0.00);
             $table->timestamps();
         });
     }
